+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Create more stack screens
 - Create welcome screen
 - Create navigator
 - Add dropdown alert manager